created a Dice class, with fields for die1 & 2, getters and setters, moved logic around
