+++ src/com/css/craps/Table.java
@@ -21,22 +21,6 @@ public class Table {
 
     }
 
-    public static int rollDice(){
-        int d1 = randomInt(1,6);
-        int d2 = randomInt(1,6);
-
-        return d1 + d2;
-    }
-    private static int randomInt(int min, int max) {
-        int result = 0;
-
-        double rand = Math.random(); // 0.0 - 0.99
-        double scaled = (rand * (max-min + 1) + min);
-        result = (int) scaled;
-
-        return result;
-    }
-
     // Method to ask user to input amount for bet
     public static int userInputBet(){
         int bet = 0;
